Working navigation using nav-* CSS properties
The focus is first given to (last) element with property nav-index:0, if any.
Navigation keys then carry the focus to indexed elements corresponding to the
values of nav-up, nav-left, nav-right and nav-down properties for up, left,
right and down keys respectively.

I put in place a logical grid for navigation, using lowercase letters a1, b1...
to differentiate it from uppercase letters used in the positionning grid.