Updated sample to use two indexed PNG with matching color palette
The images are shifted on purpose to let the background show.

Only indexed PNG can be displayed; an external color look-up table file (clut)
must be provided, which matches the colors of the palette in the PNG file.

JPEG images will be displayed at half-width, half-size, unless they are
exactly at one of the following sizes:
  * 960x540px
  * 840x520px
  * possibly others