Added styles for navigation
This stylesheet is based on experiments with nav-* properties in index.

After seeing a demonstration video of ViniSketch Designer, showing the
definition of focus as a graph, e.g. a loop for items in a list, I decided
to switch to a more limited use case: loop over items in a list.

I have also found the grid too constraining: in a page with a limited number
of links, top and left should go to previous, bottom and right should go to
next. Only when many items are included, and laid out in a grid pattern, is
the grid navigation pattern really better.

I plan to define various sets of CSS classes for lists of 1 up to 12 items:
  * link1of1
  * link1of2, link2of2
  * ...
  * link1of7, link2of7,... link7of7
  * ...
  * link1of12, link2of12,... link12of12

This should cover 80% of use cases. More complex navigation systems would be
defined with custom CSS or JavaScript.