Experiment with indexed PNG images in Gimp (ongoing)
I created a custom palette for LIME colors, and assigned it to the color map
of the indexed image. I still have issues with transparent and semi-transparent
colors: the custom palette only allows to define fully opaque colors. The
transparency is set when painting using selected color. I tried to assign the
color with the index 8 (transparency) to parts of the image, with various
levels of transparency, the color always looks opaque in LIME browser.

The same happens when opening a transparent PNG which works and saving a copy
in Gimp: the copy appears with an opaque background.