Updated text experiment to prepare the addition of gutters in layout
The padding increases the width and height of elements, in compliance with CSS.
This makes it hard to read text on the side of the screen, and on the side of
images, and on the side of adjacent text.

I will now experiment with gutters between columns to add breathing in the
grid. I am thinking of 4px gutter on top and left of each current row / column,
to let space for a border of up to 4px on each side. Borders of 4px on adjacent
cells would overlap in the gutter.