Added navigation experiment, and new navigation classes
In the new system, a link may be either an anchor (a element) or any element
with the CSS class 'link'. A numbered link class defines the order of focus,
e.g. link1, link2, link3... and by default links are expected to go down and
right. This can be customized with before-linkN and after-linkN classes, which
allows to loop from last to first and back again for example, and also expects
the next link to be right and below and the previous link up and left. It is
possible to refine the navigation further with below-linkN and above-linkN
classes, which define the navigation of the up and down arrows respectively.